requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-multipart==0.0.6
pyyaml==6.0
pillow==10.0.0
numpy==1.25.2
watchfiles==0.19.0
websockets==11.0.3
